Add debug-gui
Since a gui has a lot of dependencies, it was added as an optional feature
it is quite shitty, but that's the first step.
Currently displays the program + data stack, and the local + global scope
Added some to_syntax methods to make it easier to look at

The GUI allows one to step through the execution and see the state as it evolves.
Kinda assumed I would be able to do this with a regular debugger but mine was non-cooprative.
And this lays some groundwork for later regardless.